ai-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
langchain==0.0.267
openai==0.28.0
pinecone-client==2.2.2
weaviate-client==3.24.1
python-dotenv==1.0.0
tiktoken==0.4.0
pymongo==4.5.0
requests==2.31.0
numpy==1.25.2
pandas==2.1.0
python-multipart==0.0.6
aiohttp==3.8.5
pytest==7.4.1
pytest-asyncio==0.21.1 